make it clear that arm-2d provides the default helium acceleration for blend
